requirements.txt
mistralai>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
streamlit>=1.28.0
